requirements.txt
flask==3.0.2
PyPDF2==3.0.1
spacy==3.7.2
python-docx==1.0.1
pandas==2.1.3
scikit-learn==1.3.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
